+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Added
 * Add support for flags: pushed lines are no longer commented.
 * Add previously used aliases as hints to new timesheets.
 * Add `--today`, `--since` and `--until` support for `status` and `commit` commands.
+* Add `--used` option to `alias list` command (#89).
 
 Changed
 -------